Storage management improved, bugs fixed

+++ app/memory/useStorage.js
@@ -1,6 +1,7 @@
 import AsyncStorage from "@react-native-async-storage/async-storage";
 import { useContext, useState } from "react";
 import * as FileSystem from "expo-file-system";
+import RNFS from "react-native-fs";
 
 import asyncForEach from "../componets/scripts/asyncForEach";
 import keyfields from "./keyfields";
@@ -33,16 +34,31 @@ const getAllKeys = async () => {
   return keys;
 };
 
-const removeData = async (dataKeys, setStorageUsed, storageUsed) => {
+const removeData = async (
+  dataKeys,
+  setStorageUsed,
+  storageUsed,
+  deleteImages = false
+) => {
   //dataKeys should be an array
   const keys = [...dataKeys];
   try {
-    asyncForEach(keys, async (key) => {
-      const data = await getData(key);
-      if (data)
-        if (data[keyfields.THUMB] !== null)
-          FileSystem.deleteAsync(data[keyfields.THUMB]);
-    });
+    if (deleteImages)
+      await asyncForEach(keys, async (key) => {
+        // const data = await getData(key);
+        getData(key).then(async (data) => {
+          if (data) {
+            if (data[keyfields.THUMB] !== null)
+              FileSystem.deleteAsync(data[keyfields.THUMB]);
+
+            if (data[keyfields.IMAGES] !== null)
+              asyncForEach(data[keyfields.IMAGES], async (element) => {
+                if (element) RNFS.unlink(element);
+              });
+          }
+        });
+      });
+
     await AsyncStorage.multiRemove(keys);
     setStorageUsed(!storageUsed);
   } catch (e) {
